created javadoc, class PrintLogic
