+++ course1/semester2/hw3/task2/src/c1_s2/konovalov/task3_2/Main.java
@@ -8,7 +8,6 @@ public class Main {
         Scanner input = new Scanner(System.in);
         Printer printer = getPrinter(input);
 
-
         if (printer != null) {
             System.out.println("printer was created");
             printCommand();
