+++ course1/semester2/hw3/task2/src/c1_s2/konovalov/task3_2/printerTester.java
@@ -4,7 +4,7 @@ import org.junit.Test;
 import java.util.Random;
 import static junit.framework.TestCase.fail;
 
-public class printerTester {
+public class PrinterTester {
     @Test
     public void testConsolePrinter() {
         Printer printer = new ConsolePrinter();
